fix(review): harden lane-4 review findings
Two-agent review of the #265/#281/#295 batch:

- quit.set_grid_target_depth bounds-checks the Double before any
  Int conversion — Int(1e300) traps, so a config typo could kill
  the WM (High; the #58 lesson, matching setResizeStep).
- The remaining-gap field live-parses each keystroke: on macOS a
  button click never blurs a text field, so the commit-on-blur
  proxy let Adjust gaps read a stale value while the field showed
  the new one (Medium).
- quit.grid_target_depth clamps into range on decode (the
  clampedWidth precedent) with a pinning test; the round-trip
  suite now mutates quitGridTargetDepth so a dropped decode line
  cannot pass silently (Medium).
- BorderStyle.remainingGapRange is the shared 0-100 source for
  the command clamp and the GUI field (the targetDepthRange
  pattern).
- SettingsRows.swift split at the 350 ceiling: StepperRow and
  PlacementPicker move to their own files; RemainingGapField
  extracted from FocusBorderEditor.
- gatherWindows snapshots the density target once so frames and
  the raise circle structurally share one partition; StepperRow
  asserts the labelHidden/help exclusivity; maxDimension comment
  lists the prose mirror sites; fit_gaps docs state clamp-vs-fail.
